% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{create_dataset}
\alias{create_dataset}
\title{Create dataset from data}
\usage{
create_dataset(
  name,
  train_x,
  train_y,
  val_x = NULL,
  val_y = NULL,
  xtype,
  ytype,
  device
)
}
\arguments{
\item{name}{Name of the dataset}

\item{train_x}{A matrix for training data}

\item{train_y}{A matrix/vector for training label}

\item{val_x}{A matrix for validation data (can be NULL)}

\item{val_y}{A matrix/vector for validation label (can be NULL)}

\item{xtype}{torch_dtype of x}

\item{ytype}{torch_dtype of y}

\item{device}{torch_device}
}
\value{
A list of two datasets, "training" and "validation"
(validation is NULL if val_x or val_y is not given)
}
\description{
Create dataset from data
}
