% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\name{generate_net}
\alias{generate_net}
\title{Generate network}
\usage{
generate_net(topology, name = "defaultnet", device = NULL)
}
\arguments{
\item{topology}{Topology parameter of the network}

\item{name}{Name of the network}

\item{device}{device of the network}
}
\description{
Generate network
}
