% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpletorch.R
\name{predict.nn_module}
\alias{predict.nn_module}
\title{Predict}
\usage{
\method{predict}{nn_module}(model, x)
}
\arguments{
\item{model}{The neural network (nn_module)}

\item{x}{A matrix or torch_tensor}
}
\value{
An output matrix
}
\description{
Predict
}
