% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpletorch.R
\name{train}
\alias{train}
\title{Do training of the neural network}
\usage{
train(dl, val_dl = NULL, topology, optim, loss, nepoch, lr = 0.01)
}
\arguments{
\item{dl}{A dataloader of the training data}

\item{val_dl}{A dataloader of the validation data}

\item{topology}{Topology of the network}

\item{optim}{Name of the optimizer}

\item{loss}{Name of the loss function}

\item{nepoch}{Number of epochs}

\item{lr}{The learning rate}
}
\value{
A list of the trained model, the training loss and the validation loss
}
\description{
Do training of the neural network
}
