% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Do training of the neural network}
\usage{
train(
  dl,
  val_dl = NULL,
  topology,
  optim,
  loss,
  nepoch,
  lr = 0.01,
  save_model = FALSE,
  save_filename = "model\%d.torch",
  check = TRUE,
  verbose = TRUE,
  device = NULL
)
}
\arguments{
\item{dl}{A dataloader of the training data}

\item{val_dl}{A dataloader of the validation data}

\item{topology}{Topology of the network}

\item{optim}{Name of the optimizer}

\item{loss}{Name of the loss function}

\item{nepoch}{Number of epochs}

\item{lr}{The learning rate}

\item{save_model}{If TRUE, save the models obtained at each epoch}

\item{save_filename}{Template of the saved files. \%d is replaced with the epoch number}

\item{check}{Check the dimension of the parameter and network, type of the data, etc.}

\item{verbose}{Output the loss during the training}

\item{device}{torch_device for training}
}
\value{
A list of the trained model, the training loss and the validation loss
}
\description{
Do training of the neural network
}
